requirements.txt
python-docx
# glob